# server/test/fixtures/officepulse-openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "OfficePulse Integration API",
    "version": "1.0.0",
    "description": "Native Asterisk inventory and opt-in POC provisioning, integration call diagnostics and signed callbacks. Asterisk remains the PBX source of truth; provisioning writes its Realtime tables directly and keeps no desired-state copy. These docs are publicly readable; administrative requests remain restricted to trusted backends."
  },
  "servers": [
    {
      "url": "/",
      "description": "This OfficePulse API deployment"
    }
  ],
...
  "paths": {
...
    "/v1/admin/pbx/extensions": {
      "get": {
        "tags": ["Backend API"],
        "summary": "List native extensions and approved contexts",
        "description": "Private admitted backend network or Identity-authenticated Operations gateway with tenant authorization and mutation CSRF. PBX mutations and DID reads require PBX_PROVISIONING_ENABLED=true plus dedicated writer credentials. Disabled mutation routes are absent. Credentials are returned only by extension creation; responses must not be cached.",
        "x-access-policy": "trusted-backend-network",
        "parameters": [
          {
            "name": "iTenantId",
            "in": "query",
            "required": true,
            "description": "Exactly one canonical Identity tenant ID; duplicate values are rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped Asterisk records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionInventory"
                }
              }
            }
          },
          "403": {
            "description": "Caller not admitted or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Absent/other-tenant native object, or provisioning disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate object, remaining reference, manual route, or concurrent change; refresh and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or out-of-scope input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database, schema, ownership mapping, or provisioning unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Backend API"],
        "summary": "Create a tenant extension; disclose SIP secret once",
        "description": "Private admitted backend network or Identity-authenticated Operations gateway with tenant authorization and mutation CSRF. PBX mutations and DID reads require PBX_PROVISIONING_ENABLED=true plus dedicated writer credentials. Disabled mutation routes are absent. Credentials are returned only by extension creation; responses must not be cached.",
        "x-access-policy": "trusted-backend-network",
        "parameters": [
          {
            "name": "iTenantId",
            "in": "query",
            "required": true,
            "description": "Exactly one canonical Identity tenant ID; duplicate values are rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ExtensionCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction committed; effective Asterisk state has not been verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExtensionCreated"
                }
              }
            }
          },
          "403": {
            "description": "Caller not admitted or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Absent/other-tenant native object, or provisioning disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate object, remaining reference, manual route, or concurrent change; refresh and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or out-of-scope input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database, schema, ownership mapping, or provisioning unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/pbx/extensions/{extension}": {
      "delete": {
        "tags": ["Backend API"],
        "summary": "Delete an owned extension bundle and saved queue memberships",
        "description": "Private admitted backend network or Identity-authenticated Operations gateway with tenant authorization and mutation CSRF. PBX mutations and DID reads require PBX_PROVISIONING_ENABLED=true plus dedicated writer credentials. Disabled mutation routes are absent. Credentials are returned only by extension creation; responses must not be cached.",
        "x-access-policy": "trusted-backend-network",
        "parameters": [
          {
            "name": "iTenantId",
            "in": "query",
            "required": true,
            "description": "Exactly one canonical Identity tenant ID; duplicate values are rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "extension",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[0-9]{2,12}$"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Transaction committed; effective Asterisk state has not been verified"
          },
          "403": {
            "description": "Caller not admitted or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Absent/other-tenant native object, or provisioning disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate object, remaining reference, manual route, or concurrent change; refresh and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or out-of-scope input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database, schema, ownership mapping, or provisioning unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/pbx/queues": {
      "get": {
        "tags": ["Backend API"],
        "summary": "List explicitly owned native queues and members",
        "description": "Private admitted backend network or Identity-authenticated Operations gateway with tenant authorization and mutation CSRF. PBX mutations and DID reads require PBX_PROVISIONING_ENABLED=true plus dedicated writer credentials. Disabled mutation routes are absent. Credentials are returned only by extension creation; responses must not be cached.",
        "x-access-policy": "trusted-backend-network",
        "parameters": [
          {
            "name": "iTenantId",
            "in": "query",
            "required": true,
            "description": "Exactly one canonical Identity tenant ID; duplicate values are rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Scoped Asterisk records",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueInventory"
                }
              }
            }
          },
          "403": {
            "description": "Caller not admitted or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Absent/other-tenant native object, or provisioning disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate object, remaining reference, manual route, or concurrent change; refresh and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or out-of-scope input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database, schema, ownership mapping, or provisioning unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Backend API"],
        "summary": "Create a queue from a friendly slug (native ID tN.slug) or approved legacy name",
        "description": "Private admitted backend network or Identity-authenticated Operations gateway with tenant authorization and mutation CSRF. PBX mutations and DID reads require PBX_PROVISIONING_ENABLED=true plus dedicated writer credentials. Disabled mutation routes are absent. Credentials are returned only by extension creation; responses must not be cached.",
        "x-access-policy": "trusted-backend-network",
        "parameters": [
          {
            "name": "iTenantId",
            "in": "query",
            "required": true,
            "description": "Exactly one canonical Identity tenant ID; duplicate values are rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/QueueCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Transaction committed; effective Asterisk state has not been verified",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/QueueCreated"
                }
              }
            }
          },
          "403": {
            "description": "Caller not admitted or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Absent/other-tenant native object, or provisioning disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate object, remaining reference, manual route, or concurrent change; refresh and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or out-of-scope input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database, schema, ownership mapping, or provisioning unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/v1/admin/pbx/queues/{queue}": {
      "delete": {
        "tags": ["Backend API"],
        "summary": "Delete a queue and members; conflict while a DID references it",
        "description": "Private admitted backend network or Identity-authenticated Operations gateway with tenant authorization and mutation CSRF. PBX mutations and DID reads require PBX_PROVISIONING_ENABLED=true plus dedicated writer credentials. Disabled mutation routes are absent. Credentials are returned only by extension creation; responses must not be cached.",
        "x-access-policy": "trusted-backend-network",
        "parameters": [
          {
            "name": "iTenantId",
            "in": "query",
            "required": true,
            "description": "Exactly one canonical Identity tenant ID; duplicate values are rejected.",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "queue",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_.-]{1,80}$",
              "maxLength": 80
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Transaction committed; effective Asterisk state has not been verified"
          },
          "403": {
            "description": "Caller not admitted or not authorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "Absent/other-tenant native object, or provisioning disabled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate object, remaining reference, manual route, or concurrent change; refresh and retry",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "422": {
            "description": "Invalid fields or out-of-scope input",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "503": {
            "description": "Database, schema, ownership mapping, or provisioning unavailable",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "Error": {
        "type": "object",
        "properties": {
          "error": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "details": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "ExtensionInventory": {
        "type": "object",
        "additionalProperties": false,
        "required": ["source", "iTenantId", "provisioningEnabled", "contexts", "extensions"],
        "properties": {
          "source": {
            "type": "string",
            "enum": ["asterisk"]
          },
          "iTenantId": {
            "type": "integer"
          },
          "provisioningEnabled": {
            "type": "boolean"
          },
          "contexts": {
            "type": "array",
            "items": {
              "type": "string",
              "pattern": "^[a-zA-Z0-9_.-]{1,40}$",
              "maxLength": 40
            }
          },
          "extensions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Extension"
            }
          }
        }
      },
      "ExtensionCreate": {
        "type": "object",
        "additionalProperties": false,
        "required": ["extension"],
        "properties": {
          "extension": {
            "type": "string",
            "pattern": "^[0-9]{2,12}$"
          },
          "context": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_.-]{1,40}$",
            "maxLength": 40,
            "description": "Required when more than one approved context exists."
          },
          "displayName": {
            "type": "string",
            "minLength": 1,
            "maxLength": 33,
            "description": "No quotes, angle brackets, backslashes or control characters. The formatted caller ID must fit the installed 40-character column."
          },
          "callerIdNumber": {
            "type": "string",
            "pattern": "^\\+[1-9][0-9]{6,14}$",
            "example": "+19496501147"
          }
        }
      },
      "ExtensionCreated": {
        "type": "object",
        "additionalProperties": false,
        "required": ["extension", "sipUsername", "sipSecret", "applyState"],
        "properties": {
          "extension": {
            "type": "string"
          },
          "sipUsername": {
            "type": "string"
          },
          "sipSecret": {
            "type": "string",
            "description": "One-time disclosure. Never logged, cached, persisted by AidaAdmin, or returned by inventory/replays."
          },
          "applyState": {
            "type": "string",
            "enum": ["committed", "active", "unknown"],
            "description": "This implementation returns committed after a transaction and unknown for native inventory. It does not verify effective Asterisk state or return active."
          }
        }
      },
...
      "QueueInventory": {
        "type": "object",
        "additionalProperties": false,
        "required": ["source", "iTenantId", "provisioningEnabled", "queues"],
        "properties": {
          "source": {
            "type": "string",
            "enum": ["asterisk"]
          },
          "iTenantId": {
            "type": "integer"
          },
          "provisioningEnabled": {
            "type": "boolean"
          },
          "queues": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Queue"
            }
          }
        }
      },
      "QueueCreate": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name"],
        "properties": {
          "name": {
            "type": "string",
            "pattern": "^[a-zA-Z0-9_.-]{1,80}$",
            "maxLength": 80,
            "description": "Friendly slug up to 60 characters, or exact approved legacy name up to 80. Generated native ID is tN.slug."
          },
          "strategy": {
            "type": "string",
            "enum": [
              "ringall",
              "leastrecent",
              "fewestcalls",
              "random",
              "rrmemory",
              "linear",
              "wrandom"
            ],
            "default": "ringall"
          }
        }
      },
      "QueueCreated": {
        "type": "object",
        "additionalProperties": false,
        "required": ["name", "strategy", "applyState"],
        "properties": {
          "name": {
            "type": "string"
          },
          "strategy": {
            "type": "string",
            "enum": [
              "ringall",
              "leastrecent",
              "fewestcalls",
              "random",
              "rrmemory",
              "linear",
              "wrandom"
            ]
          },
          "applyState": {
            "type": "string",
            "enum": ["committed", "active", "unknown"],
            "description": "This implementation returns committed after a transaction and unknown for native inventory. It does not verify effective Asterisk state or return active."
          }
        }
      },
...
    }
  }
}